+++ rockylinux-server-postinstall.sh
@@ -3,7 +3,7 @@
 # 0/1 : enable cockpit
 cockpit=1
 # List of additional software to install	
-addsoftwares=(vim tar git tmux ncdu htop rsync)
+addsoftwares=(vim tar git tmux ncdu htop rsync git-extras)
 # 0/1 : Désactiver le parefeu firewalld
 disablefirewalld=0
 # enforcing/permissive/disabled : Statut de SELinux à activer
@@ -16,14 +16,21 @@ if [[ $EUID -ne 0 ]]
 then
 	sudo chmod +x "$(dirname "$0")/$0"
 	sudo "$(dirname "$0")/$0"
-	exit;
+	exit
 fi
 
 # SELinux
 sed -e "s/SELINUX=.*/SELINUX=$selinux/" -i /etc/sysconfig/selinux
 
 # Upgrade
-dnf -y --nogpgcheck --refresh upgrade
+dnf upgrade --refresh --nogpgcheck
+dnf check
+dnf autoremove
+dnf install fwupd -y
+fwupdmgr get-devices
+fwupdmgr refresh --force
+fwupdmgr get-updates
+fwupdmgr update
 
 # Turn on EPEL repo
 dnf -y install epel-release &&  dnf repolist
@@ -33,8 +40,9 @@ if [[ -n ${addsoftwares[*]} ]]
 then
 	dnf install --nogpgcheck -y "${addsoftwares[@]}"
 fi
-type -p tmux >/dev/null && curl -JLO https://raw.githubusercontent.com/imomaliev/tmux-bash-completion/master/completions/tmux && mv tmux /usr/share/bash-completion/completions/tmux
-
+if [ -f /usr/share/bash-completion/completions/tmux ]; then
+	curl -JLO https://raw.githubusercontent.com/imomaliev/tmux-bash-completion/master/completions/tmux && mv tmux /usr/share/bash-completion/completions/tmux
+fi
 
 # Cockpit
 
@@ -47,6 +55,12 @@ then
 	firewall-cmd --add-service=cockpit --permanent
 	firewall-cmd --reload
 fi
+# Git exstra and toolbelt
+if [ ! -f /usr/local/bin/git-cleave ]; then
+	git clone --depth 1 --single-branch --branch v1.7.0 https://github.com/nvie/git-toolbelt > /dev/null
+	cp git-toolbelt/git-* /usr/local/bin
+	rm -r git-toolbelt
+fi
 
 
 # Disable firewalld
@@ -58,26 +72,29 @@ then
 fi
 
 # install or update starship
-curl -fsS https://starship.rs/install.sh | bash -s -- -y
+type -p starship >/dev/null || curl -fsS https://starship.rs/install.sh | bash -s -- -y >/dev/null
 
 
 # install rust
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+type -p ruspup >/dev/null || curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
 
 # install go
-case $arch in
-	aarch64) 
-		curl -JLO https://golang.org/dl/go1.17.2.linux-arm64.tar.gz
-		rm -rf /usr/local/go && tar -C /usr/local -xzf go1.17.2.linux-arm64.tar.gz
-		rm go1.17.2.linux-arm64.tar.gz
+if [ ! -f /usr/local/go/bin/go ]; then
+	goVersion='1.17.2'
+	case $arch in
+	aarch64)
+		curl -JLO https://golang.org/dl/go${goVersion}.linux-arm64.tar.gz
+		rm -rf /usr/local/go && tar -C /usr/local -xzf go${goVersion}.linux-arm64.tar.gz
+		rm go${goVersion}.linux-arm64.tar.gz
 		;;
-	x86_64) 
-	    curl -JLO https://golang.org/dl/go1.17.2.linux-amd64.tar.gz
-		rm -rf /usr/local/go && tar -C /usr/local -xzf go1.17.2.linux-amd64.tar.gz
-		rm go1.17.2.linux-amd64.tar.gz
+	x86_64)
+		curl -JLO https://golang.org/dl/go${goVersion}.linux-amd64.tar.gz
+		rm -rf /usr/local/go && tar -C /usr/local -xzf go${goVersion}.linux-amd64.tar.gz
+		rm go${goVersion}.linux-amd64.tar.gz
 		;;
-	*);;
-esac
-echo "Add /usr/local/go/bin to the PATH environment variable"
+	*) ;;
+	esac
+	echo "Add /usr/local/go/bin to the PATH environment variable"
+fi
 
 echo "Preparation completed, it is recommended to restart!"